+++ docs/source/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.12.8 (2026-01-DD)
+- Fix #315: DuplicateNameError in case a task returns several ExternalTableReference objects with the same table name
+- Implement #316: Support top level import pydiverse.pipedag.CacheValidationMode
+
 ## 0.12.7 (2025-12-15)
 - Fix: cache invalidation works more reliably for pydiverse.transform and ibis
 - Fix: be more fault-tolerant in duckdb view parsing (ParquetTableStore)